CLI/BUG Return zero value from seed interface
Fix a bug where, after successfully generating new input files
using unique random seeds, a non-zero exit value would be returned.
This often indicates a failure and did not represent the success
of the interface.

Now, a zero value is returned and, if the user has not specified
quiet, the generated files are printed to the screen on unique lines.
This can be used to quickly write to a file, e.g.

$ python -m serpentTools seed file N > inputs.txt