CLI Clean up to-matlab interface - verbosity support
If the user has requested quiet, then nothing is printed
from the interface. Otherwise, one of two actions is performed.
If the user has not specified any level of verbosity, then
just the name of the output file is generated. If some level
of verbosity is given, the an info message is given with the
input file and the generated output file