DEV Tweak str2vec to accept single word, use map
Closes #305 by adding a little extra logic into the conversion.
If there are no spaces in the input string, then the input
string is directly passed into the output type, e.g. to make a
list or of a single string. Otherwise, the previous logic prevails.
Add an additional unit test for this specific test case, and for
reading a result file with ADFs.

Utilize python map function and numpy fromiter to gain
some performace improvements.