API Add __getitem__ and __setitem__ for HomogUniv
Allows these to like a dictionary, as these are convinience
methods to access **JUST THE EXPECTED VALUES** of requested
data. e.g.
>>> h = HomogUniv("setget", None, None, None)
>>> h['infDummy'] = 1.0
>>> h.infExp
{'infDummy': 1.0}
>>> h.infUnc:
{}
>>> h['infDummy']
1.0

These methods do not perform conversions between
SERPENT_CASE and mixedCase, but will make the
placement of the data funky. The quantity INF_TOT
would be placed in the gc dictionary, due to the lookup checks.