API Add threshold argument for Detector.meshPlot
Uses the numpy masked arrays to create a copy of the
plot data, where all values less than or equal to the
threshold are "masked". Matplotlib knows how to handle
these arrays, and will not apply a color to these
sections of the data.

Updated some documentation on the method to reflect this
change.