TST Better legend testing for DepletedMaterial plot
If only a single isotope is plotted, and legend != True, then
the plot method will not draw a legend, causing no labels
to be created. The function getLegendTexts returns an
empty list in this case. Therefore, one depletion plot test
compares against an un-labeled plot, while the formatted
plot enforces the legend construction.