API/DEV Variable-specific type conversion for ResultsReader
When reading, the ResultsReader creates a lookup dictionary used
to convert SERPENT universe variables to usable forms. This is done
first on a type basis: values that have a single value, no uncertianty,
of a specific type. Example: DF_SYM [int], DF_VOLUME [float]. Then,
there are variables that should be arrays of floats without
uncertainties such as MICRO_NG, DF_SURF_AREA. Variables with no
conversion, DF_SURFACE, are also allowed. The converter is expected
to return two items, (value[s], uncertainty[ies]). The uncertainties
can be None, and these value can be arrays or single quantities.
Some checks are also included for values like INF_KEFF which should
be floats, but we want to store k and it's uncertainty as floats,
not as single-item arrays.
Through populating universe data, the converter function for
each variable is added to a lookup dictionary.

This allows more flexible storage of variable data, as some quantities
in SERPENT result files are provided just a strings, while some are
vectors without uncertainties. The issue this is addressing is how
some discontinuity factor data is stored. Prior to this commit,
the vector of surface areas would be split and partioned off as if it
had uncertainties. This is not consistent with how the data is presented
in the output file, and has been mitigated with this commit.

The ADF test has been updated and expanded to check for types of
variables and for their contents.