API All universe objects from coe files are strings, not int
This removes the workarounds introduced for compatibility in
PR #321 / Issue #318, where it was pointed out that Serpent allows
universe names to be generic strings rather than integers.
The workaround allowed the user to request universe ``0`` from the
BranchCollector and BranchContainer objects, with a warning
that this feature would be removed in future versions. Well that
version is 0.8.0 and the time has come.