API Refactor detectors: Import from serpentTools.detectors
Heavily refactored the detector objects. A summary of the changes
is presented here, with justifications to follow:

- Detector scores are removed - closing GH Issue #327
- Detector.fromTallyBins classmethod replaces creating a detector
  and then adding tally data in separate method
- Detector objects are more flexible, relying on properties
  to access and modify tallies, errors, and grids
- DetectorBase is removed. Detector is a fully functional class
- All detectors can be imported from serpentTools.detectors and are
  included in the main API, e.g. serpentTools.Detector
- Importing detectors from serpentTools.objects.detectors is
  deprecated, favoring serpentTools.detectors
- Heavily re-document the detector interface. The main documentation
  is now explicitly written, with no programmatic doc-string building.
  That got ugly to read.

                               NON-FIXES

- Samplers need to be fixed and have not been modified
- No checks are currently performed on the shape of incoming
  tally and error data, e.g. they could potentially be of different
  shapes

The primary justification for this change was to allow the detectors
more flexibility over how their data was provided. The previous method
relied exclusively on the addTallyData method to perform the reshaping.
Now, the user can create a Detector already knowing the, or without
knowing, the tally data. Setting tallies and errors are now done through
properties, ensuring that newly set data is of identical shape.

This goes to resolve GH issue #337 by allowing the user more natural
control over tally data. The new pytest-style test file
tests/test_detectors.py modifies the tally and error data in-place
to emulate some scaling.

Work still needs to be done to fix the sampler module, as the
sampled detectors rely upon the now removed DetectorBase