API Refactor serpentTools.samplers.SampledDetector
Similar to changes introduced in
5c9c46faf3c56658fff9806b7f7c98035488271b, the SampledDetector
is now more like a standalone object. It can be created directly
by passing allTallies and allErrors, or with an iterable of
similar detectors through SampledDetector.fromDetectors.
The implementation is a lot cleaner, relying less on hidden methods
for finalizing and computing errors. The propagation of uncertainty
is done during the init method, as is the averaging of the tally data.
Additionally, a new attribute is provided: deviation. This attribute
provides the standard deviation in the expected tally data.