API Detector.indexes is now tuple of strings
Same goes to SampledDetectors

More concise and harder to break due to immutability of
tuples. Detector.indexes now a property that checks that
tallies and/or errors are set up and the number of items
corresponds to the dimensionality of the data.

Using this tuple, the plot routines are modified to ensure
that tally data and indexes are set up prior to any slicing and
plotting.