API Use UnivTuple for keys in ResultsReader.universes
Closes part of #338 by creating a UnivTuple namedtuple that
contains universe, burnup, step, and burnup [days] to represent
a specific set of universe data. This is preferable to the
plain tuple approach because data can be accessed as attributes as
well as position. Consider the following::

>>> p = NamedTuple("0", 1, 2, 3)
>>> p[0]  # universe
"0"
>>> p.universe
"0"
>>> p.step == p[2]
True

It is preferable to use the attribute-based approach, as this
is more likely to be preserved through potential future changes.

Taught relevant aspects of the ResultsReader about this tuple,
and updated documentation