API BranchContainer is a dictionary of UnivTuple -> HomogUniv
Previously, the BranchContainer was able to store universes with
time indicated with burnup or days. In order to use a named-tuple
like object, but retaining the same API, two tuples would have to
be created. One with (universe, burnup, step) and the other
with (universe, day, step). Rather than go down this convoluted route,
the UnivTuple is used. This will cause some breaking changes, as
the key ["0", 0, 0] no longer exists in the dictionary. Instead,
a key ["0", 0, 0, None] or ["0", None, 0, 0] exists. The
getUniv method still works without any external changes.

Another change introduced here is to treat the BranchContainer
as a subclass of dictionary, explicitly map UnivTuples to
HomogUniv instances. The universes attribute is still provided as a
property, but returns the instance itself.

Related: GH Issue #338