TST Add results test for multiple gcu, no burnup
In support of verifying GH Issue #345, a simple test has been
added. This test builds a valid result file using the
InnerAssembly_res.m data file, but without any burnup. This file is
chosen because it contains three sets of homogenized universes.
Regular expressions are used to exclude any burnup data. By
counting the number of "% Increase counter" lines, the new
file is truncated after writing data for all universes at
step 0.

A new file is created in a fixture and yielded to the underlying
test. The test performs a check that only three universes are
present, and then checks that burnup, step, and days for all
universes are set to zero.

The test currently fails, as this is the point of this bug fix.