BUG Correctly get variable groups for version 2.1.31
The lack of a 2.1.31 block in variables.yaml meant that if the
user specified
```
rc["serpentVersion"] = "2.1.31"  # also the default!!
rc["variableGroups"] = ["six-ff", "n-balance"]
```
the six factor formula information would not be expanded. Furthermore,
since 2.1.31 and the base [2.1.29] differ in the neutron balance data,
some information would be missing.

This commit adds the 2.1.30 group as an anchor in variables.yaml
that 2.1.31 references. This means we don't have to re-write the
whole block. Did have to change 2.1.30 -> 2-1-30 in the file as
only alphanumeric characters are allowed for yaml anchors.

A test has been added for this expansion, and the variable group
documentation has been updated.