API Allow modifying aesthetics of depletion spread plot
Users can pass sampleKwargs and/or meanKwargs to alter the
look of the sampled data or mean data, respectively. Unless
provided, the sampled data are plotted in semi-transparent
black with the mean data in a soft blue.
These must be dictionaries of matplotlib acceptable
quantities, e.g. ``{"c": "k", "alpha": 0.5}``