DEV Use list of arrays to improve results parsing
Introduced a helper class, ListOfArrays, to improve performance of
the ResultsReader.  Instead of using ``numpy.vstack`` each time a
new row is appended into the ``ResultsReader.resdata``, the row is
simply appended into a list. Some checking is done to ensure that
the rows have consistent shape and data type.

After reading, the final arrays are constructed and placed back
in the ``resdata`` dictionary. This approach improves performance
and scalability, especially for longer rows, e.g. more burnup steps.