BUG Do not process arrays from sensitivity histories
This closes GH issue #356 but changing how arrays are processed
in the sensitivity reader. The variable names are processed using
a single for-loop rather than multiple index calls. The new routine
also reads the _N_GEN information to be used later in storing
these arrays. Examination of these arrays can help the determination
of ideal number of latent generations, as described in the Serpent
Wiki using the sens opt history setting.