DEV Update parsing engines- stream based
From commit 2f893ffb89bf952df3cda4b3d9d11a06b5ac3070 of the
drewtils python project - release 0.2.0rc0. The main changes are
that the KeywordParser and PatternReader now operate on readable
buffers, not files to be opened. This means the
```
        with KeywordParser(...) as parser:
                <actions>
```
syntax is no longer supported. The motivation for this change
is in support of allowing parsers and files to also act on streams,
not strictly file names. This can allow the parsers to utilize
in-memory buffers like ``io.StringIO``, or, potentially, files that
have been compressed using the ``zipfile`` library.

Most of our readers work internally on the streams, because the
opened file is what we care about. But the changes may or may not
(likely the latter) be straightforward to implement. Thankfully this
branch / feature is all about breaking the file objects from the
readers so there already is some internal changes coming.