DEV Use pyproject.toml to force building with setuptools
PEP 518 introduced the pyproject.toml file for containing
the bare minimum of build requirements. Some packages that use
cython or f2py or numpy in the build process include those. We
don't so our pyproject.toml file is very minimal. In doing this, we
are ensured to work in a build environment that has setuptools,
removing some ugly logic from the setup file.

The toml file is kind of like an INI file, and is described
more through the links in PEP 518

- https://www.python.org/dev/peps/pep-0518/
- https://github.com/toml-lang/toml

This file has also been included in our MANIFEST file to ensure
it is shipped with source distributions