DEV Remove automatic logging configuration
This tends to have really bad down-stream effects if someone
is developing with serpentTools. Ideally, the logging level,
handles, and general configuration should be up to the end user
or the developer. The configuration in serpentTools.messages has been
removed, leaving only a logger that can be easily found with
``logging.getLogger("serpentTools")``

Some reading that inspired this change:
https://docs.python.org/3/howto/logging.html#configuring-logging-for-a-library