DEV Improve flexibility of formatPlot, placeLegend
Use Axes.set for a single call in formatPlot - fix #379

Developers can now provide additional keyword arguments for
legends and for the plot to both of these methods that will
be forwarded directly along to Axes.set and Axes.legend.

Changed the ncol argument in formatPlot to be legendcols to
be more explicit, as all other options are applied to the Axes
instance.

These changes were propagated across the classes that rely on
formatPlot, which turned out to be quite a few.