BUG Put branching fractions on XSData
Fixes GH #387 by placing the branching fractions in a dedicated
XSData instance in xsections, rather than in metadata. Since the
line that signals the start of the branching data begins with
i<iso>_<lib>_mt, a new XSData entry is added to xsections. But,
when the bra_f data arrives, the intended target does not exist
in metadata.

Now, the target dictionary is xsections, giving a proper XSData
instance that can plot energy-dependent branching ratios.