DOC More explicit special and inherited members
Document specifically the len, iter, contains, get/set item methods
for all classes in the project. Additionally, use the inherited-members
autodoc option to show inherited members. Looking at the current
build of the CartesianDetector documentation
https://serpent-tools.readthedocs.io/en/0.9.2/generated/serpentTools.CartesianDetector.html
the plot and slice methods are not documented since they don't exist
directly on the CartesianDetector, but on the base Detector.

Additionally, the inherited-members option supports skipping
members that would be inherited from specific classes. This is used
to stop the searching at our NamedObject and BaseObject classes.
Otherwise a lot of un-needed and confusing members are documented.