API Dictionary-like access for HistoryReader
Much of this already existed, but this changes improves the
get and iter methods. Special methods are also placed near the
top of the class declaration as well.

The history reader test has been updated to test the contains,
get, __getitem__, iter, and items methods. The getitem test
has been merged into the special methods test