DOC Fix README.rst: bad links, remove Zenodo link
The zenodo section could be confusing, because it may lead people
to think they should cite using the zenodo DOI, rather than the
one provided by Nuclear Science and Engineering. Since we have
the publication, it's time to use it.

Also fixed some bad links / structure that caused pypi to reject
the distribution, as the description would not be rendered
property as RST