DEV Introduce next SerpentFile ABC
The new class resides can be imported as
serpentTools.next.base.SerpentFile, and
describes the interface for the serpentTools.next design. Concrete
classes are expected to provide a method
for reading a Serpent source file,
presented as an IOBase. The abstract class
handles converting strings or pathlib.Path
instances to readable streams