API Provide next Detector{File,Reader}
The new classes are contained in
serpentTools/next/detector.py and are fully implemented.
The DetectorFile inherits from a dictionary and SerpentFile,
enforcing that all values in the dictionary are Detector instances.
The __setitem__ and update methods are over-written to back this
up.

The dedicated DetectorReader is intended more for repeated
use, and can be created externally to the DetectorFile.
The main benefit is likely to be experienced if many files
are to be read, and if a subset of detectors are to be
processed from those files. A set of strings is used to
control what detectors are processed from the file, defaulting
to all detectors.