DEV str2vec only converts string -> array
Other cases are less likely across this package, and can be taken care
of usually with some flavor of list comprehension. This removes a lot
of logic inside this function, which is used a lot throughout this
project. In most cases, we could probably remove the function as it is
a one-liner now.

Tests in test_utils.py relating to this function have been removed as
they involved a lot of boilerplate for setting up and testing the now
defunct modes for str2vec. Instead, a simple pytest test has been
written that creates a random vector of three sizes (mainly to see how
it handles vectors of a single entry), converts to a space-separated
string, and then back to vector.

In a couple cases across the project, the use of this function
has been changed to just use list-comprehension