BUG Support more values like pspec on xsplot reader
xsplot objects now have a `misc` dictionary that contains data
we don't store directly on attributes. The primary use for this is
reading photon line spectrum, `pspec` entries from xsplot files.
These fields can be controlled by the `_misc` "private" attribute
on the xsplotreader if more rapid turnaround is needed by users.

Tests have been added to demonstrate that the reader is reading
these data and that a warning is raised for unidentified fields.

Closes #454